+++ zerver/views/user_groups.py
@@ -57,6 +57,7 @@ def add_user_group(
     name: str,
     members: Json[list[int]],
     description: str,
+    can_join_group: Json[int | AnonymousSettingGroupDict] | None = None,
     can_manage_group: Json[int | AnonymousSettingGroupDict] | None = None,
     can_mention_group: Json[int | AnonymousSettingGroupDict] | None = None,
 ) -> HttpResponse:
@@ -116,12 +117,14 @@ def edit_user_group(
     user_group_id: PathOnly[int],
     name: str | None = None,
     description: str | None = None,
+    can_join_group: Json[GroupSettingChangeRequest] | None = None,
     can_manage_group: Json[GroupSettingChangeRequest] | None = None,
     can_mention_group: Json[GroupSettingChangeRequest] | None = None,
 ) -> HttpResponse:
     if (
         name is None
         and description is None
+        and can_join_group is None
         and can_manage_group is None
         and can_mention_group is None
     ):
@@ -132,7 +135,10 @@ def edit_user_group(
     )
 
     if user_group.deactivated and (
-        description is not None or can_mention_group is not None or can_manage_group is not None
+        description is not None
+        or can_join_group is not None
+        or can_mention_group is not None
+        or can_manage_group is not None
     ):
         raise JsonableError(_("You can only change name of deactivated user groups"))